requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
cffi==1.17.1
cryptography==44.0.0
fastapi==0.115.6
greenlet==3.1.1
idna==3.10
passlib==1.7.4
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-status==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
